pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mcp-git"
version = "0.1.0"
description = "Git operations MCP server for AI Agents"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Kirky.X"}
]
keywords = ["git", "mcp", "ai-agent", "model-context-protocol"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # MCP Protocol
    "mcp>=0.9.0",

    # Git Operations
    "GitPython>=3.1.40",

    # Async
    "anyio>=4.0.0",
    "asyncio-throttle>=1.0.2",

    # Database
    "aiosqlite>=0.19.0",

    # Configuration
    "pydantic>=2.5.0",
    "python-dotenv>=1.0.0",

    # Security
    "secrecy>=1.2.0",

    # Logging
    "loguru>=0.7.0",

    # Utilities
    "uuid6>=2024.1.12",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
    "pre-commit>=3.6.0",
]
test = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "coverage[toml]>=7.3.0",
]
docs = [
    "sphinx>=7.2.0",
    "sphinx-rtd-theme>=1.3.0",
    "sphinx-autodoc-typehints>=1.24.0",
]
security = [
    "safety>=2.3.0",
    "bandit>=1.7.0",
    "pip-audit>=2.6.0",
]

[project.urls]
Homepage = "https://github.com/Kirky-X/mcp-git"
Documentation = "https://mcp-git.readthedocs.io"
Repository = "https://github.com/Kirky-X/mcp-git"
"Bug Tracker" = "https://github.com/Kirky-X/mcp-git/issues"

[project.scripts]
mcp-git = "mcp_git.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["mcp_git*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.black]
line-length = 100
target-version = ['py310']
include = '\.pyi?$'

[tool.ruff]
line-length = 100
target-version = "py310"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501", # line too long
    "B008", # do not perform function calls in argument defaults
]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

# Coverage 配置
[tool.coverage.run]
source = ["mcp_git"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__pycache__/*",
    "*/site-packages/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "raise AssertionError",
    "raise NotImplementedError",
    "@abstractmethod",
    "@overload",
]
precision = 2
show_missing = true

[tool.coverage.html]
directory = "htmlcov"

# Bandit 配置（安全扫描）
[tool.bandit]
exclude_dirs = ["/tests", "/test", "test_*.py"]
skips = ["B101", "B601"]
